split out AKS E2E source files
